Altered Frontend appearance and functionality
Removed the navbar from all pages and replaced the functionality inside
the sidebar.

Created new JavaScript functions to alter the appearance of
the pages in the frontend.

Minor Change to the backend to respond with the equipment_types.

todo: create the to normalise the equipment widths and set the
equipment_types value into Session in the backend.

 On branch main
 Your branch is up to date with 'origin/main'.

 Changes to be committed:
	modified:   public_html/backend/controlers/equipment/equipment_controler.php
	modified:   public_html/frontend/css/iframes/equipment/tabs/yur_eq.css
	modified:   public_html/frontend/css/sidebar/sidebar.css
	modified:   public_html/frontend/html/sidebar/sidebar.php
	modified:   public_html/frontend/iframes/equipment/tabs/yur_eq.html
	modified:   public_html/frontend/js/iframes/equipment/tab_controler.js
	new file:   public_html/frontend/js/sidebar/sidebar.js