Functional Equipment Delete
Verified if the code works as intended for deletion in the db. Had to
remove the foreign key constraint "equipment_id" from equipment log
otherwise the equipment would not be deleted.

Created a function that validates if the users request references any
users items and or equipment in the database

 On branch main
 Your branch is up to date with 'origin/main'.

 Changes to be committed:
	modified:   backend/controllers/request/equipment/request.php
	modified:   backend/controllers/request/equipment/request_authentication.php
	modified:   backend/controllers/request/equipment/request_sanitization.php
	modified:   backend/controllers/request/equipment/request_validation.php
	modified:   backend/crud/common/common_functions.php
	modified:   backend/crud/common/query_generators.php
	modified:   backend/crud/delete/equipment_delete.php
	modified:   backend/crud/read/equipment_query.php